+++ src/Adapter/InMemory.php
@@ -5,6 +5,8 @@ namespace Innmind\Filesystem\Adapter;
 
 use Innmind\Filesystem\{
     Adapter,
+    Adapter\InMemory\Overwrite,
+    Adapter\InMemory\Merge,
     File,
     Name,
     Directory,
@@ -18,20 +20,27 @@ use Innmind\Immutable\{
 final class InMemory implements Adapter
 {
     private Directory $root;
+    private Overwrite|Merge $behaviour;
 
-    private function __construct()
+    private function __construct(Overwrite|Merge $behaviour)
     {
         $this->root = Directory\Directory::named('root');
+        $this->behaviour = $behaviour;
     }
 
     public static function new(): self
     {
-        return new self;
+        return new self(new Overwrite);
+    }
+
+    public static function emulateFilesystem(): self
+    {
+        return new self(new Merge);
     }
 
     public function add(File $file): void
     {
-        $this->root = $this->root->add($file);
+        $this->root = ($this->behaviour)($this->root, $file);
     }
 
     public function get(Name $file): Maybe